Moved constants and functions related to projects from common_storage.ts to project.ts.
