+++ src/storage/create_python_files.ts
@@ -24,10 +24,10 @@
 import * as Blockly from 'blockly/core';
 import { extendedPythonGenerator } from '../editor/extended_python_generator';
 import {
-    Project,
     Storage,
     parseModuleContentText } from './common_storage';
 import { Module } from './module';
+import { Project } from './project';
 import { pascalCaseToSnakeCase } from './names';
 import JSZip from 'jszip';
 import { GeneratorContext } from '../editor/generator_context';
